bug fix: show correct value (brightness) on lovelace light card
Because the light card has no max value attribute (is hardcoded 255) we need a to add a max attribute taken from the object to calculate the correct value.

This bug fix works only for manual created entities at the moment.
For auto detect we need a max value, but dont know how the detection works in detail